[link] refactor(text-button-mapper): refactored to enable flow
