+++ packages/plugin-link/src/createLinkPlugin.js
@@ -3,14 +3,14 @@ import { createBasePlugin, decorateComponentWithProps } from 'wix-rich-content-c
 import { EXTERNAL_LINK_TYPE } from './types';
 import { Strategy, Component } from './LinkDecorator';
 import styles from '../statics/link-viewer.scss';
-import createLinkToolbar from './toolbar/createLinkToolbar';
+import { linkTextButtonMapper } from './toolbar/LinkTextButtonMapper';
 
 const createLinkPlugin = (config = {}) => {
   const type = EXTERNAL_LINK_TYPE;
   const { decorator, helpers, theme, isMobile, t, anchorTarget, relValue, getEditorState, setEditorState } = config;
   const plugin = createLinkifyPlugin({ target: anchorTarget, rel: relValue, theme: theme || styles });
 
-  const toolbar = createLinkToolbar({ helpers, theme, isMobile, t, anchorTarget, relValue, getEditorState, setEditorState });
+  const toolbar = linkTextButtonMapper({ helpers, theme, isMobile, t, anchorTarget, relValue, getEditorState, setEditorState });
 
   plugin.decorators.push({
     strategy: Strategy,
